pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "genai-faq"
version = "0.1.0"
description = "Production-ready RAG-powered FAQ system"
requires-python = ">=3.11,<3.13"
dependencies = [
    "langgraph>=0.2",
    "langchain>=0.3",
    "langchain-openai>=0.2",
    "langchain-community>=0.3",
    "chromadb>=0.5",
    "openai>=1.50",
    "sentence-transformers>=3.0",
    "ragas>=0.2",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "streamlit>=1.38",
    "structlog>=24.0",
    "python-multipart>=0.0.9",
    "slowapi>=0.1",
    "pypdf>=4.0",
    "beautifulsoup4>=4.12",
    "httpx>=0.27",
    "pydantic-settings>=2.0",
    "lxml>=5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.6",
]

[tool.setuptools.packages.find]
include = ["app*", "ui*", "scripts*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"
